fix: gp typing again

+++ common/gradient_penalty.py
@@ -35,12 +35,11 @@ def gradient_penalty(model: nn.Module, data: th.Tensor, k: float = 1,) -> th.Ten
     gradients = th.autograd.grad(outputs=prediction[1], inputs=data,
                                  grad_outputs=th.ones(prediction[1].size()),
                                  create_graph=True, retain_graph=True)[0]
-    gradients = gradients.to(th.float)
 
     gradients = gradients.view(data.shape[0], -1)
     gradients_norm = th.sqrt(th.sum(gradients ** 2, dim=1) + 1e-12)
 
-    return ((gradients_norm - k) ** 2).mean()
+    return ((gradients_norm - k) ** 2).mean().to(th.float)
 
 
 def gradient_penalty_for_continues_critic(network: nn.Module,
@@ -52,4 +51,4 @@ def gradient_penalty_for_continues_critic(network: nn.Module,
     with th.set_grad_enabled(not share_feature_extractor):
         features = features_extractor(observations)
     qvalue_input = th.cat([features, actions], dim=1)
-    return gradient_penalty(network, qvalue_input, k)
+    return gradient_penalty(network, qvalue_input, k).to(th.float)
